+++ op-batcher/compressor/compressors.go
@@ -1,7 +1,7 @@
 package compressor
 
 import (
-	"github.com/ethereum-optimism/optimism/op-node/rollup/derive/compression"
+	"github.com/ethereum-optimism/optimism/op-node/rollup/compression"
 )
 
 type FactoryFunc func(Config) (compression.Compressor, error)
