+++ op-node/rollup/compression/channel_compressor.go
